requirements.txt
requests>=2.31.0
discord.py>=2.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
aiohttp>=3.8.0
pytz>=2023.3
